Module hey_listen::sync

The parallel/async dispatcher module.

Re-exports

pub use async_dispatcher::AsyncDispatcher;

pub use parallel_dispatcher::ParallelDispatcher;

pub use priority_dispatcher::PriorityDispatcher;

Modules

async_dispatcher
This module contains the async dispatcher.

parallel_dispatcher
This module contains the parallel dispatcher.

priority_dispatcher
This module contains the priority dispatcher.

Enums

AsyncDispatchResult
An enum returning a request from a Listener to its async event-dispatcher.

ParallelDispatchResult
An enum returning a request from a Listener to its parallel event-dispatcher.

PriorityDispatcherResult
An enum returning a request from a listener to its sync event-dispatcher. This enum is not restricted to dispatcher residing in the sync-module. A request will be processed by the event-dispatcher depending on the variant:

Traits

AsyncListener
Every event-receiver needs to implement this trait in order to receive dispatched events. T being the type you use for events, e.g. an Enum.

Listener
Every event-receiver needs to implement this trait in order to receive dispatched events. T being the type you use for events, e.g. an Enum.

ParallelListener
Every event-receiver needs to implement this trait in order to receive dispatched events. T being the type you use for events, e.g. an Enum.

PriorityListener
Every event-receiver needs to implement this trait in order to receive dispatched events. T being the type you use for events, e.g. an Enum.