Struct hey_listen::sync::priority_dispatcher::PriorityDispatcher

pub struct PriorityDispatcher<P, T> where
    P: Ord,
    T: PartialEq + Eq + Hash + Clone + Send + Sync + 'static,  { /* fields omitted */ }

In charge of prioritised sync dispatching to all listeners. Opposed to EventListener, this structure utilises one BTreeMap per event-type to order listeners by a given priority-level.

Note: Consider implementing your own Ord-trait, if you want a different priority.

Implementations

impl<P, T> PriorityDispatcher<P, T> where
P: Ord + Clone,
T: PartialEq + Eq + Hash + Clone + Send + Sync + 'static,

pub fn add_listener<D: PriorityListener<T> + Send + Sync + 'static>(
&mut self,
event_key: T,
listener: D,
priority: P
)

Adds a Listener to listen for an event_identifier, considering a given priority implementing the Ord-trait, to sort dispatch-order. If event_identifier is a new HashMap-key, it will be added.

Note: If your Enum owns fields you need to consider implementing the Hash- and PartialEq-trait if you want to ignore fields, see second example for an implementation-suggestion.

Examples

Adding a Listener to the dispatcher:

use std::sync::Arc;
use hey_listen::{
   RwLock,
   sync::{PriorityListener, PriorityDispatcher, PriorityDispatcherRequest},
};

#[derive(Clone, Eq, Hash, PartialEq)]
enum Event {
    EventType,
}

struct ListenerStruct;

impl PriorityListener<Event> for ListenerStruct {
    fn on_event(&self, event: &Event) -> Option<PriorityDispatcherRequest> { None }
}

fn main() {
    let listener = ListenerStruct;
    let mut dispatcher: PriorityDispatcher<u32, Event> = PriorityDispatcher::default();

    dispatcher.add_listener(Event::EventType, listener, 1);
}

Declaring your own Hash- and PartialEq-trait to bypass hashing on fields:

use std::hash::{Hash, Hasher};
use std::mem::discriminant;

#[derive(Clone)]
enum Event {
    TestVariant(i32),
}

impl Hash for Event {
    fn hash<H: Hasher>(&self, _state: &mut H) {}
}

impl PartialEq for Event {
    fn eq(&self, other: &Event) -> bool {
        discriminant(self) == discriminant(other)
    }
}

impl Eq for Event {}

pub fn dispatch_event(&mut self, event_identifier: &T)

All Listeners listening to a passed event_identifier will be called via their implemented on_event-method. Fns returning Result with Ok(()) will be retained and Err(SyncDispatchResult::StopListening) will cause them to be removed from the event-dispatcher.

Notice: Listeners will called ordered by their priority-level.